+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - fix: Make `pulumi login` respect configuration in `Pulumi.yaml`
   ([#1299](https://github.com/pulumi/actions/pull/1299))
 - feat: Strip ansi control characters from GitHub step Summary and PR comment
-  ([]())
+  ([#1312](https://github.com/pulumi/actions/pull/1312))
 
 ---