+++ CHANGELOG.md
@@ -2,16 +2,21 @@
 
 ## HEAD (Unreleased)
 
+**(none)**
+
+---
+
+## 6.0.1 (2025-01-13)
+
 - fix: Make `pulumi login` respect configuration in `Pulumi.yaml`
   ([#1299](https://github.com/pulumi/actions/pull/1299))
+
 - fix: Log stderr from commands
   ([#1316](https://github.com/pulumi/actions/issues/1316))
 
 - fix: Make the `output` command work without having to be in a working
   directory that contains a Pulumi.yaml file.
   ([#1327](https://github.com/pulumi/actions/pull/1327))
-
----
 
 ## 6.0.0 (2024-09-27)